Monitoring: Merge reducers/ui.ts into components/monitoring/utils.ts
